+++ clang/test/SemaCXX/constant-expression-cxx14.cpp
@@ -1456,3 +1456,29 @@ constexpr bool missingCase() {
     1u: return false; // expected-error {{expected 'case' keyword before expression}}
   }
 }
+
+namespace GH183290 {
+struct A {
+  constexpr char a() const { return 'Z'; }
+  char b = a();
+};
+
+// expected-error@+1 {{expected class name}}
+struct B : sizeof(int[c]) {};
+
+struct C {
+  B b;
+  // expected-note@+1 {{overridden virtual function is here}}
+  virtual const A *d() const;
+};
+
+struct D : C {
+  // expected-error@+1 {{return type of virtual function 'd' is not covariant with the return type of the function it overrides ('const B *' is not derived from 'const A *')}}
+  constexpr virtual const B *d() const { return &this->b; }
+};
+
+constexpr D e;
+constexpr const C *f = &e;
+// expected-error@+1 {{static assertion expression is not an integral constant expression}}
+static_assert(f->d()->b == 'Z', "");
+}